+++ src/extension.ts
@@ -57,6 +57,7 @@ import { EditPublicIdentitiesConfigCommand } from './commands/tenant/editPublicI
 import { EditAccessRequestConfigCommand } from './commands/tenant/editAccessRequestConfigCommand';
 import { NewAttributeSearchConfigCommand } from './commands/NewAttributeSearchConfigCommand';
 import { EditPasswordConfigCommand } from './commands/tenant/editPasswordConfigCommand';
+import { GenerateDigitTokenCommand } from './commands/tenant/generateDigitTokenCommand';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -101,11 +102,18 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand(commands.EDIT_ACCESS_REQUEST_CONFIG,
 			editAccessRequestConfigCommand.execute,
 			editAccessRequestConfigCommand));
+	
 	const editPasswordConfigCommand = new EditPasswordConfigCommand()
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.EDIT_PASSWORD_ORG_CONFIG,
 			editPasswordConfigCommand.execute,
 			editPasswordConfigCommand));
+	
+	const generateDigitTokenCommand = new GenerateDigitTokenCommand(tenantService)
+	context.subscriptions.push(
+		vscode.commands.registerCommand(commands.GENERATE_DIGIT_TOKEN,
+			generateDigitTokenCommand.execute,
+			generateDigitTokenCommand));
 
 	const identityNowDataProvider = new IdentityNowDataProvider(context, tenantService);
 	vscode.window.registerTreeDataProvider(commands.TREE_VIEW, identityNowDataProvider);