+++ src/wizard/inputIdentityQueryStep.ts
@@ -2,11 +2,11 @@ import { requiredValidator } from "../validator/requiredValidator";
 import { InputPromptStep } from "./inputPromptStep";
 import { WizardContext } from "./wizardContext";
 
-export class InputOwnerStep extends InputPromptStep<WizardContext> {
-    constructor() {
+export class InputIdentityQueryStep extends InputPromptStep<WizardContext> {
+    constructor(name = "ownerQuery", displayName = "owner",) {
         super({
-            name: "ownerQuery",
-            displayName: "owner",
+            name,
+            displayName,
             options: {
                 validateInput: (s: string) => { return requiredValidator.validate(s); },
                 learnMoreLink: "https://documentation.sailpoint.com/saas/help/search/searchable-fields.html#searching-identity-data"